backend/requirements.txt
uvicorn>=0.18.3
pydantic[dotenv]>=1.9.0
aiosqlite>=0.17.0
fastapi-amis-admin>=0.5.0,<0.6.0
fastapi-user-auth>=0.5.0,<0.6.0
fastapi-scheduler>=0.0.6,<0.1.0
